requirements-deployment.txt
valohai-utils
tensorflow==2.4.1
Pillow
python-multipart